src/test/java/uk/co/joecastle/msscbeerservice/controller/BeerControllerTest.java
package uk.co.joecastle.msscbeerservice.controller;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import uk.co.joecastle.msscbeerservice.model.BeerModel;

import java.util.UUID;

import static org.junit.jupiter.api.Assertions.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(BeerController.class)
class BeerControllerTest {

    @Autowired
    MockMvc mockMvc;

    @Autowired
    ObjectMapper mapper;

    @Test
    void getBeerById() throws Exception {
        mockMvc.perform(get("/api/v1/beer/" + UUID.randomUUID())
                .accept(MediaType.APPLICATION_JSON))
            .andExpect(status().isOk());
    }

    @Test
    void saveNewBeer() throws Exception {
        BeerModel beerModel = BeerModel.builder().id(UUID.randomUUID()).build();;
        String beerModelToJson = mapper.writeValueAsString(beerModel);

        mockMvc.perform(post("/api/v1/beer")
                .contentType(MediaType.APPLICATION_JSON)
                .content(beerModelToJson))
            .andExpect(status().isCreated())
            .andExpect(header().string("Location", "http://localhost/api/v1/beer/" + beerModel.getId()));
    }

    @Test
    void updateBeerById() throws Exception {
        BeerModel beerModel = BeerModel.builder().build();;
        String beerModelToJson = mapper.writeValueAsString(beerModel);

        mockMvc.perform(put("/api/v1/beer/" + UUID.randomUUID())
                .contentType(MediaType.APPLICATION_JSON)
                .content(beerModelToJson))
            .andExpect(status().isNoContent());
    }
}